Added progress bars
Added a file to contain progress bars for the fit() function.
Also made sure that everything can work together when algorithms are stacked on top of eachother, with the exception of BERT Theseus because that changes the archetecture.
That has caused me to make some fixes to BERT_Theseus and ThiNet.