TOFD RUNS*!
It appears to run sucessfully but I don't have any idea how well it is transferring the knowledge. Also the auxillary classifiers are not structured as in the paper, I will modify them to match next. But I wanted to get something working even if it is not a 100% implementation