BERT_Theseus multi-select
I noticed that the BERT-Theseus code could not be run on multiple sub-modules at the same time so I have updated it to work.

Also added an error catch block to ThiNet because it was sometimes getting a "Singular Matrix" error for Linear algebra, I do not know why that is happening.