Added updated results files +changesnges
As the title says, the results files have gotten updated versions (In particular ADMM and TOFD)

There is now a way to load a specific config row by command line.

A couple incorrect default config settings were returned.

Graphing has been updated to use the new data

DAIS has a remove function that is more technically correct (it should function the same as before, just be more correct to what is literally written)

And stopped the annoying Note 16 from occuring every load, now it just occurs when it was supposed to.