Data issues
ThiNet and BERT Theseus were accessing the training and validation data for training. I feel so embarrassed about that.

Also added a new dataset, the ACI flows data from the ACI data repository.

In addition I am including the ADMM test that shows that the method of adding the v layers I am using is equivalent in backpropigation to the actual method used in the code example (The difference is because the code example was too rigid for our tests.)