Update algorithmruns.py
For some reason DAIS was only training for 10 epochs? That could be a problem

Also, recreation run was limited to only pruning to 50% for some reason? i think it was originally to match the distribution of the rest of the algorithms but it looks fine when I remove the cap.