minimize requirements, single written config
Minimized the requirements file (there were many lines that were not required, and I think there are even more)

Made it possible to have config values that can only be set once, this is for config values that are more meta (such as number of runs)

also split main.py into functions. And added a couple more descriptions for some functions.