Update LRP.py
Found it!

The problem was that the R was being split over the biases as well (the paper says to do that) which was a sink for the relevance, but since we weighed positive and negative relevance the same, the bias could sink a lot more relevance than we expected as well as sinking negative relevance.